# api/openapi.yaml
openapi: 3.1.0
info:
    title: PP Prime Construction API
    version: 0.1.0
    description: |
        สัญญา API ระหว่างหน้าบ้าน (Angular) และหลังบ้านของระบบบริหารงานก่อสร้าง PP Prime Construction

        หลักการ
        - เงินเป็นบาท (number) ทศนิยมได้ 2 ตำแหน่ง
        - วันที่ตามแผนใช้ `date` (YYYY-MM-DD) เวลาที่เกิดเหตุการณ์ใช้ `date-time` (ISO 8601 พร้อมเขตเวลา)
        - สถานะ/ประเภทเป็นรหัสภาษาอังกฤษ หน้าบ้านแปลเป็นข้อความไทยเอง
        - ข้อผิดพลาดตอบเป็น `application/problem+json` (RFC 9457)
        - การคำนวณทางธุรกิจ (ระดับผู้อนุมัติ กำไร-ขาดทุน ไทม์ไลน์ สถานะงวด) ทำที่หลังบ้าน
        - ทุกการเปลี่ยนแปลงข้อมูล หลังบ้านต้องบันทึก Audit Log เอง (หน้าบ้านไม่มี API เขียน Audit Log)
        - สิทธิ์การเข้าถึงตรวจที่หลังบ้านตามบทบาทใน `GET /roles`; ข้อมูลที่บทบาทนั้นไม่มีสิทธิ์เห็นต้องไม่ถูกส่งมา
servers:
    - url: /api
      description: ค่าเริ่มต้น (หน้าบ้านตั้ง apiBaseUrl ได้)
security:
    - bearerAuth: []
...
paths:
...
    /house-plans/{code}:
        parameters:
            - { name: code, in: path, required: true, schema: { type: string } }
        get:
            tags: [projects]
            summary: แบบบ้าน แปลนรายชั้น และไฟล์โมเดล 3D
            operationId: getHousePlan
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/HousePlan' }
                '404': { $ref: '#/components/responses/NotFound' }
    /personnel:
        get:
            tags: [personnel]
            operationId: listPersonnel
            parameters:
                - { name: q, in: query, schema: { type: string } }
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: '#/components/schemas/Personnel' }
        post:
            tags: [personnel]
            summary: เพิ่มบุคลากร (หลังบ้านออกรหัสบุคลากร)
            operationId: createPersonnel
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: '#/components/schemas/PersonnelInput' }
            responses:
                '201':
                    description: Created
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/Personnel' }
                '422': { $ref: '#/components/responses/ValidationError' }
    /personnel/{id}:
        parameters:
            - { name: id, in: path, required: true, schema: { type: string } }
        get:
            tags: [personnel]
            operationId: getPersonnel
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/Personnel' }
                '404': { $ref: '#/components/responses/NotFound' }
        put:
            tags: [personnel]
            operationId: updatePersonnel
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: '#/components/schemas/PersonnelInput' }
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/Personnel' }
                '404': { $ref: '#/components/responses/NotFound' }
                '422': { $ref: '#/components/responses/ValidationError' }
    /personnel/license-alerts:
        get:
            tags: [personnel]
            summary: ใบอนุญาต/เอกสารที่หมดอายุหรือใกล้หมดอายุ
            operationId: listLicenseAlerts
            parameters:
                - { name: withinDays, in: query, schema: { type: integer, default: 90, minimum: 1 } }
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: '#/components/schemas/LicenseAlert' }
...
    /audit-logs:
        get:
            tags: [audit]
            summary: บันทึกการกระทำ (ล่าสุดก่อน, อ่านอย่างเดียว)
            operationId: listAuditLogs
            parameters:
                - { name: module, in: query, schema: { $ref: '#/components/schemas/AuditModule' } }
                - { name: userId, in: query, schema: { type: string } }
                - { name: q, in: query, description: ค้นหาการกระทำ รายการ รายละเอียด, schema: { type: string } }
                - { name: from, in: query, schema: { type: string, format: date-time } }
                - { name: to, in: query, schema: { type: string, format: date-time } }
                - $ref: '#/components/parameters/Page'
                - $ref: '#/components/parameters/PageSize'
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/AuditLogPage' }
                '403': { $ref: '#/components/responses/Forbidden' }
    /audit-logs/export:
        get:
            tags: [audit]
            summary: ส่งออก CSV ตามตัวกรองเดียวกับ /audit-logs (UTF-8 พร้อม BOM)
            operationId: exportAuditLogs
            parameters:
                - { name: module, in: query, schema: { $ref: '#/components/schemas/AuditModule' } }
                - { name: userId, in: query, schema: { type: string } }
                - { name: q, in: query, schema: { type: string } }
            responses:
                '200':
                    description: ไฟล์ CSV
                    content:
                        text/csv:
                            schema: { type: string }
                '403': { $ref: '#/components/responses/Forbidden' }
    /dashboard/summary:
        get:
            tags: [dashboard]
            summary: กำไร-ขาดทุนรายโครงการและยอดรวมบริษัท (เจ้าของบริษัท)
            operationId: getDashboardSummary
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/DashboardSummary' }
                '403': { $ref: '#/components/responses/Forbidden' }
...
    /dashboard/cash-flow:
        get:
            tags: [dashboard]
            summary: เงินรับ-เงินจ่ายรายเดือน (เดือนล่าสุดนับถึงวันนี้)
            operationId: getCashFlow
            parameters:
                - { name: months, in: query, schema: { type: integer, default: 6, minimum: 1, maximum: 24 } }
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: '#/components/schemas/MonthlyCashFlow' }
                '403': { $ref: '#/components/responses/Forbidden' }
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    parameters:
...
        Page:
            name: page
            in: query
            description: เริ่มที่ 1
            schema: { type: integer, minimum: 1, default: 1 }
        PageSize:
            name: pageSize
            in: query
            schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    responses:
...
        Forbidden:
            description: บทบาทนี้ไม่มีสิทธิ์
            content:
                application/problem+json:
                    schema: { $ref: '#/components/schemas/Problem' }
        NotFound:
            description: ไม่พบข้อมูล
            content:
                application/problem+json:
                    schema: { $ref: '#/components/schemas/Problem' }
...
        ValidationError:
            description: ข้อมูลไม่ถูกต้อง
            content:
                application/problem+json:
                    schema: { $ref: '#/components/schemas/Problem' }
    schemas:
        Problem:
            type: object
            required: [title, status]
            properties:
                type: { type: string, format: uri, default: about:blank }
                title: { type: string, description: ข้อความสั้นสำหรับแสดงผู้ใช้ (ภาษาไทย) }
                status: { type: integer }
                detail: { type: string }
                errors:
                    type: object
                    description: ข้อผิดพลาดรายช่อง (ชื่อช่อง → ข้อความ)
                    additionalProperties: { type: string }
        PageMeta:
            type: object
            required: [page, pageSize, total]
            properties:
                page: { type: integer }
                pageSize: { type: integer }
                total: { type: integer }

        # ---------- ผู้ใช้และสิทธิ์ ----------
...
        UserRef:
            type: object
            required: [id, name, roleLabel]
            properties:
                id: { type: string }
                name: { type: string }
                roleLabel: { type: string, description: ชื่อบทบาท ณ เวลาที่เกิดเหตุการณ์ }
...
        RoomKind:
            type: string
            enum: [bedroom, bathroom, living, dining, kitchen, garage, stair, void, other]
        HousePlan:
            type: object
            required: [code, name, description, width, depth, floorHeight, parking, floors]
            properties:
                code: { type: string }
                name: { type: string }
                description: { type: string }
                width: { type: number, description: เมตร }
                depth: { type: number, description: เมตร }
                floorHeight: { type: number }
                parking: { type: integer }
                renderUrl: { type: string, format: uri, description: ภาพทัศนียภาพ }
                modelUrl: { type: string, format: uri, description: ไฟล์โมเดล .glb }
                floors:
                    type: array
                    items:
                        type: object
                        required: [label, rooms]
                        properties:
                            label: { type: string }
                            rooms:
                                type: array
                                items:
                                    type: object
                                    required: [name, kind, x, y, w, h]
                                    properties:
                                        name: { type: string }
                                        kind: { $ref: '#/components/schemas/RoomKind' }
                                        x: { type: number }
                                        y: { type: number }
                                        w: { type: number }
                                        h: { type: number }

        # ---------- บุคลากร ----------
        PersonnelDocument:
            type: object
            required: [type, fileName, url]
            properties:
                type: { type: string, description: ประเภทเอกสาร เช่น สำเนาบัตรประชาชน }
                fileName: { type: string }
                url: { type: string, format: uri, description: ลิงก์จาก POST /uploads }
                expiresAt: { type: string, format: date }
                uploadedBy: { type: string, description: หลังบ้านใส่ให้ (ไม่ต้องส่ง) }
        PersonnelTraining:
            type: object
            required: [name, completedAt]
            properties:
                name: { type: string }
                completedAt: { type: string, format: date }
                expiresAt: { type: string, format: date }
        PersonnelSensitive:
            type: object
            description: ข้อมูลอ่อนไหว ส่งมาเฉพาะบทบาทที่มีสิทธิ์ (เจ้าของบริษัท บัญชี/การเงิน) บทบาทอื่นไม่มีฟิลด์นี้ และค่าที่ส่งเข้ามาจะถูกละเว้น
            properties:
                compensationType: { type: string, description: เงินเดือน / รายวัน / เหมางาน }
                compensationRate: { type: number }
                compensationHistory: { type: string }
                withholdingTaxTerms: { type: string }
                bankName: { type: string }
                bankAccountNumber: { type: string }
                bankAccountName: { type: string }
                taxpayerId: { type: string }
                withholdingCertificate: { type: string }
                performanceRating: { type: string }
                performanceNotes: { type: string }
        PersonnelInput:
            type: object
            description: ข้อมูลที่แก้ไขได้ในหน้าบุคลากร (วันที่ว่างให้ละฟิลด์ไว้)
            required: [fullName]
            properties:
                photoUrl: { type: string, description: ลิงก์จาก POST /uploads }
                fullName: { type: string, minLength: 1 }
                nationalId: { type: string }
                position: { type: string }
                birthDate: { type: string, format: date }
                hireDate: { type: string, format: date }
                endDate: { type: string, format: date }
                phone: { type: string }
                lineId: { type: string }
                email: { type: string }
                registeredAddress: { type: string }
                currentAddress: { type: string }
                emergencyContactName: { type: string }
                emergencyContactRelation: { type: string }
                emergencyContactPhone: { type: string }
                employmentType: { type: string, description: ประจำ / รายเดือน / รายวัน }
                department: { type: string }
                supervisor: { type: string }
                expertise: { type: string }
                workAreas: { type: string }
                availability: { type: string }
                education: { type: string }
                institution: { type: string }
                otherCertificates: { type: string }
                professionalLicenseNumber: { type: string }
                professionalLicenseType: { type: string }
                professionalLicenseLevel: { type: string }
                professionalLicenseExpiresAt: { type: string, format: date }
                workPermitExpiresAt: { type: string, format: date }
                socialSecurityNumber: { type: string }
                issuedEquipment: { type: string }
                documents:
                    type: array
                    items: { $ref: '#/components/schemas/PersonnelDocument' }
                trainings:
                    type: array
                    items: { $ref: '#/components/schemas/PersonnelTraining' }
                sensitive: { $ref: '#/components/schemas/PersonnelSensitive' }
        Personnel:
            allOf:
                - $ref: '#/components/schemas/PersonnelInput'
                - type: object
                  required: [id, employeeCode, projectHistory, auditHistory]
                  properties:
                      id: { type: string }
                      employeeCode: { type: string, description: หลังบ้านออกให้เมื่อเพิ่มบุคลากร, examples: [EMP690001] }
                      projectHistory:
                          type: array
                          description: หลังบ้านสร้างจากการมอบหมายงาน (อ่านอย่างเดียว)
                          items:
                              type: object
                              required: [projectCode, role, period, assignment, dailyReport]
                              properties:
                                  projectCode: { type: string }
                                  role: { type: string }
                                  period: { type: string }
                                  assignment: { type: string }
                                  dailyReport: { type: string }
                                  paidAmount: { type: number, description: ส่งมาเฉพาะบทบาทที่เห็นข้อมูลอ่อนไหว }
                      auditHistory:
                          type: array
                          description: ประวัติการแก้ไข (อ่านอย่างเดียว)
                          items:
                              type: object
                              required: [at, userName, action]
                              properties:
                                  at: { type: string, format: date-time }
                                  userName: { type: string }
                                  action: { type: string }
        LicenseAlert:
            type: object
            required: [personnelId, fullName, license, expiresAt, expired]
            properties:
                personnelId: { type: string }
                fullName: { type: string }
                license: { type: string }
                expiresAt: { type: string, format: date }
                expired: { type: boolean }

        # ---------- การอนุมัติ ----------
...
        AuditModule:
            type: string
            enum: [approval, project, personnel, procurement, finance, system]
        AuditEntry:
            type: object
            required: [id, at, user, module, action, target]
            properties:
                id: { type: string }
                at: { type: string, format: date-time }
                user: { $ref: '#/components/schemas/UserRef' }
                module: { $ref: '#/components/schemas/AuditModule' }
                action: { type: string, description: ข้อความการกระทำ (ภาษาไทย) }
                target: { type: string, description: เลขที่/รหัสของสิ่งที่ถูกกระทำ }
                detail: { type: string }
        AuditLogPage:
            allOf:
                - $ref: '#/components/schemas/PageMeta'
                - type: object
                  required: [items]
                  properties:
                      items:
                          type: array
                          items: { $ref: '#/components/schemas/AuditEntry' }

        # ---------- Dashboard ----------
        ProjectHealth:
            type: string
            enum: [normal, behind, over-budget, loss]
        ProjectFinance:
            type: object
            required: [projectCode, customerName, contractValue, progress, plannedProgress, budgetCost, earnedRevenue, actualCost, grossProfit, margin, forecastProfit, cashReceived, receivable, health]
            properties:
                projectCode: { type: string }
                customerName: { type: string }
                contractValue: { type: number }
                progress: { type: integer }
                plannedProgress: { type: integer, description: ความคืบหน้าที่ควรเป็นตามเวลา }
                budgetCost: { type: number }
                earnedRevenue: { type: number, description: รายได้รับรู้ตามความคืบหน้า }
                actualCost: { type: number }
                grossProfit: { type: number }
                margin: { type: number, description: '% ของรายได้รับรู้' }
                forecastProfit: { type: number, description: กำไรที่คาดเมื่อจบโครงการ }
                cashReceived: { type: number }
                receivable: { type: number }
                health: { $ref: '#/components/schemas/ProjectHealth' }
...
        DashboardSummary:
            type: object
            required: [asOf, totals, projects]
            properties:
                asOf: { type: string, format: date-time }
                totals:
                    type: object
                    required: [contractValue, backlog, activeProjects, earnedRevenue, actualCost, grossProfit, margin, forecastProfit, cashReceived, receivable]
                    properties:
                        contractValue: { type: number }
                        backlog: { type: number }
                        activeProjects: { type: integer }
                        earnedRevenue: { type: number }
                        actualCost: { type: number }
                        grossProfit: { type: number }
                        margin: { type: number }
                        forecastProfit: { type: number }
                        cashReceived: { type: number }
                        receivable: { type: number }
                projects:
                    type: array
                    items: { $ref: '#/components/schemas/ProjectFinance' }
        MonthlyCashFlow:
            type: object
            required: [month, cashIn, cashOut]
            properties:
                month: { type: string, description: YYYY-MM, examples: ['2026-10'] }
                cashIn: { type: number }
                cashOut: { type: number }